Rename: SearchBar 컴포넌트 위치 이동 - 홈 페이지에서 재사용되기 때문에 공용 컴포넌트 폴더로 이동
