+++ src/app/components/SearchBar.tsx
@@ -12,10 +12,10 @@ export default function SearchBar() {
   };
   return (
     <form
-      onSubmit={handleSubmit}
       role="search"
       aria-label="도서 검색"
-      className="mb-14 flex w-96 max-w-md rounded-lg bg-[#fefefe] p-4"
+      onSubmit={handleSubmit}
+      className="mb-14 flex w-2/6 min-w-72 rounded-lg bg-[#fefefe] px-4 py-5 shadow-sm"
     >
       <label htmlFor="searchInput" className="sr-only">
         검색
@@ -27,7 +27,7 @@ export default function SearchBar() {
         id="searchInput"
         placeholder="Search..."
         aria-describedby="search-hint"
-        className="grow text-font-textPrimary outline-none"
+        className="grow text-lg text-font-textPrimary outline-none"
         onChange={(e) => setQuery(e.target.value)}
       />
       <button type="submit" aria-label="검색 버튼" className="search">
